+++ src/index.html
@@ -31,5 +31,7 @@
 
       <input type="reset" class="clear-list" value="Clear all completed"/>
     </form>
+
+    <script type="module" src="./index.js"></script>
   </body>
 </html>
